Limit prefetch count for rabbitmq msg

+++ api/src/rabbitmq/billing.rs
@@ -663,7 +663,7 @@ pub(super) async fn process_request(
 				}
 				TotalAmount::NeedToPay(need_to_pay) => (
 					need_to_pay,
-					max(0, total_bill.total_charge - need_to_pay),
+					max(0, total_bill.total_charge as i64 - need_to_pay as i64) as u64,
 					0,
 				),
 			};
